cambios generales

+++ app/Http/Controllers/usuarioController.php
@@ -9,6 +9,8 @@ use App\Http\Requests\EditPregRequest;
 use Auth;
 use Session;
 use DB;
+use Cache;
+
 class usuarioController extends Controller
 {
     public $repositori = null;
@@ -21,35 +23,30 @@ class usuarioController extends Controller
     }
     public function create($grupo)
     {
-        return $this->repositori->create($grupo);
+        if(auth()->user()) return $this->repositori->create($grupo);
+        else return $this->repositori->index($grupo);
     }
     public function usuario()
     {
-        return response()->json(Auth::user());
+        if(! Cache::has('usuario')){
+            Cache::forever('usuario', Auth::user());
+        }
+        return response()->json(Cache::get('usuario'));
     }
     public function grupos()
     {
-        return response()->json(DB::table('grupos')->orderBy('grupo')->get());
+        if(! Cache::has('grupos')){
+            Cache::forever('grupos', DB::table('grupos')->orderBy('grupo')->get());
+        }
+        return response()->json(Cache::get('grupos'));
     }
-    public function mylike($id_pregunta)
+    public function likes($id_pregunta)
     {
-        $datos = DB::table('likes')->where('id_pregunta',$id_pregunta)->where('id_usuario',Auth::user()->id)
-        ->count();
-        return response()->json($datos);
+        return $this->repositori->likes($id_pregunta);
     }
-    public function mylike_comentarios($id_comentario)
+    public function likes_comentarios($datos)
     {
-        $datos = DB::table('like_comentario')->where('id_comentario',$id_comentario)->where('id_user',Auth::user()->id)
-        ->count();
-        return response()->json($datos);
-    }
-    public function likes($datos,$like)
-    {
-        return $this->repositori->likes($datos,$like);
-    }
-    public function likes_comentarios($datos,$like)
-    {
-        return $this->repositori->likes_comentarios($datos,$like);
+        return $this->repositori->likes_comentarios($datos);
     }
     public function notificaciones(){
         $datos = DB::table('notificaciones')->where('id_admin',Auth::user()->id)
@@ -57,12 +54,6 @@ class usuarioController extends Controller
         ->orderBy('id','desc')->get();
         return json_encode($datos);
     }
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(preguntaRequest $request)
     {
         return $this->repositori->store($request);
@@ -72,16 +63,20 @@ class usuarioController extends Controller
        return $this->repositori->editar_pregunta($request,$id);
     }
     public function eliminar_pregunta($id){
+        $preguntas = DB::table('preguntas')->where('id',$id)->where('id_usuario',Auth::user()->id)->count();
+        if($preguntas > 0)return response()->json([], 500);
+        Cache::pull('mis_preguntas');
         DB::table('preguntas')->where('id',$id)->where('id_usuario',Auth::user()->id)->delete();
+        DB::table('notificaciones')->where("id_pregunta",$id)->delete();
+        DB::table('likes')->where("id_pregunta",$id)->delete();
+        DB::table('like_comentario')->where("id_pregunta",$id)->delete();
+        DB::table('comentarios')->where("id_pregunta",$id)->delete();
         return "Exito";
     }
     public function comentarios($id)
     {
-        $datos = DB::table('comentarios')->where('id_pregunta',$id)
-        ->join('users','users.id','=','comentarios.id_usuario')
-        ->select('comentarios.id','comentario','id_usuario','id_admin','id_pregunta','likes',
-        'nombre_apellido')->get();
-        return json_encode($datos);
+        if(auth()->user()) return $this->repositori->comentarios($id);
+        else return $this->repositori->comentarios_init($id);
     }
     public function guardar_comentarios(Request $request,$id){
         return $this->repositori->guardar_comentarios($request,$id);
@@ -101,6 +96,7 @@ class usuarioController extends Controller
     public function logout()
     {
         session()->flush();
+        Cache::flush();
         Auth::logout();
         return redirect('/');
     }
